requirements.txt
botocore
pyyaml
boto3
kix

aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-route53
aws-cdk.aws-s3
aws-cdk.aws-cloudfront
aws-cdk.aws-route53-targets
aws-cdk.aws-s3-deployment
aws-cdk.aws-dynamodb
aws-cdk.aws-iam
aws-cdk.custom-resources